Reworked the section on line continuation.
Moved most general line-continuation advice to a subsection of the main
Style and Conventions section. Also reworked advice to make it more
consistent and general. Some toying with code examples.

Also changed code example environment so it does not break across pages.
Code examples are not meant to be very long (longer examples can either
use floats or a breakable version), and the page breaks visually disrupt
the layout of the code that the example was supposed to demonstrate in
the first place.

Updated PDF.